Fix: replace the using of google json library with fastjson. Format application.yml Downgrade fastjson version to 1.2.88.

+++ src/main/kotlin/org/rainark/whuassist/util/Movie.java
@@ -24,7 +24,7 @@ public class Movie {
     public String name;
     public String info;
     public String description;
-    public float rank;
+    public double rank;
     public String detailPage;
     /**
      * Get the poster image of the movie.
